Added my edits of the bot.
Fixed a rare issue where if a rolled card was owned by another user and didn't spawn any kakera reacts, the code would try to index the null array and stop working. 
Added extra functionality with auto-usage of commands like $us 20, $daily, $dk, etc. ($daily and $dk are sent once every hour to be safe, which is technically redundant, but it works).